Update the Player and Projectile scripts
The funcion HitByProjectile was renamed to hit, and the parameters are changed, if the player die, he are deactivated and the death effect is instantiated.
The projectile now call the hit funcion passin the damage and the projectile position, this position is used to calculate the angle of the death effect.